+++ mux/mux.go
@@ -18,6 +18,7 @@ import (
 	"mellium.im/xmlstream"
 	"mellium.im/xmpp"
 	"mellium.im/xmpp/disco/info"
+	"mellium.im/xmpp/disco/items"
 	"mellium.im/xmpp/internal/ns"
 	"mellium.im/xmpp/stanza"
 )
@@ -208,6 +209,43 @@ func (m *ServeMux) PresenceHandler(typ stanza.PresenceType, payload xml.Name) (h
 func (m *ServeMux) HandleXMPP(t xmlstream.TokenReadEncoder, start *xml.StartElement) error {
 	h, _ := m.Handler(start.Name)
 	return h.HandleXMPP(t, start)
+}
+
+// ForItems implements items.Iter for the mux by iterating over all child items.
+func (m *ServeMux) ForItems(node string, f func(items.Item) error) error {
+	for _, h := range m.patterns {
+		if itemIter, ok := h.(items.Iter); ok {
+			err := itemIter.ForItems(node, f)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	for _, h := range m.iqPatterns {
+		if itemIter, ok := h.(items.Iter); ok {
+			err := itemIter.ForItems(node, f)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	for _, h := range m.msgPatterns {
+		if itemIter, ok := h.(items.Iter); ok {
+			err := itemIter.ForItems(node, f)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	for _, h := range m.presencePatterns {
+		if itemIter, ok := h.(items.Iter); ok {
+			err := itemIter.ForItems(node, f)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	return nil
 }
 
 // ForFeatures implements info.FeatureIter for the mux by iterating over all